+++ inventory/projects.php
@@ -1,31 +1,61 @@
 <?php
 	include_once("../include/aperetiv.inc.php");
 
-	$customer	= new Customer($cid);
-	$project	= new Project($pid);
+	$customer	= new Customer($cid, $_PJ_auth);
+
+	if(isset($pid)) {
+		$project = new Project($customer, $_PJ_auth, $pid);
+	}
 
 	if(isset($eid)) {
-		$effort = new Effort($eid);
+		$effort = new Effort($eid, $_PJ_auth);
 	}
 
 	$center_template	= "inventory/project";
 	if(isset($new)) {
+		if(!$customer->checkUserAccess('new')) {
+			$error_message		= $GLOBALS['_PJ_strings']['error_access'];
+			include("$_PJ_root/templates/error.ihtml");
+			include_once("$_PJ_include_path/degestiv.inc.php");
+			exit;
+		}
 		$center_title		= $GLOBALS['_PJ_strings']['inventory'] . ': ' . $GLOBALS['_PJ_strings']['new_project'];
 		include("$_PJ_root/templates/add.ihtml");
 		exit;
 	}
-
 	if(isset($edit)) {
+		if($pid && !$project->checkUserAccess('write')) {
+			$error_message		= $GLOBALS['_PJ_strings']['error_access'];
+			include("$_PJ_root/templates/error.ihtml");
+			include_once("$_PJ_include_path/degestiv.inc.php");
+			exit;
+		}
 		if(isset($altered)) {
 			$data = array();
 			$data['id']							= $pid;
 			$data['customer_id']				= $cid;
-			$data['project_name']				= $project_name;
-			$data['project_desc']				= $project_desc;
-			$data['project_budget']				= $project_budget;
+			$data['project_name']				= addslashes($project_name);
+			$data['project_desc']				= addslashes($project_desc);
+			$data['project_budget']				= intval($project_budget);
 			$data['project_budget_currency']	= $project_budget_currency;
 			$data['closed']						= $closed;
-			$project = new Project($data);
+			$data['gid']						= $gid;
+			$data['user']						= $user;
+			$data['access']						= $access_owner . $access_group . $access_world;
+
+			if($data['gid'] == '') {
+				$data['gid']	= $project->giveValue('gid');
+			}
+			if($data['access'] == '') {
+				$data['access'] = $project->giveValue('access');
+			}
+			if($data['user'] == '') {
+				$data['user']	= $project->giveValue('user');
+			}
+			if($data['user'] == '') {
+				$data['user']	= $_PJ_auth->giveValue('id');
+			}
+			$project = new Project($customer, $_PJ_auth, $data);
 			$project->save();
 			$list = 1;
 		} else {
@@ -37,6 +67,12 @@
 	}
 
 	if(isset($delete) && !isset($cancel)) {
+		if(!$project->checkUserAccess('write')) {
+			$error_message		= $GLOBALS['_PJ_strings']['error_access'];
+			include("$_PJ_root/templates/error.ihtml");
+			include_once("$_PJ_include_path/degestiv.inc.php");
+			exit;
+		}
 		if(isset($confirm)) {
 			$project->delete();
 			$list = 1;
@@ -47,10 +83,16 @@
 		}
 	}
 
-	if($expand) {
-		$efforts = new EffortList($pid);
+	if($cid && !$customer->checkUserAccess('read')) {
+		$error_message		= $GLOBALS['_PJ_strings']['error_access'];
+		include("$_PJ_root/templates/error.ihtml");
+		include_once("$_PJ_include_path/degestiv.inc.php");
+		exit;
 	}
-	$projects = new ProjectList($cid, $shown['cp']);
+	if($expand) {
+		$efforts = new EffortList($customer, $project, $_PJ_auth);
+	}
+	$projects = new ProjectList($customer, $_PJ_auth, $shown['cp']);
 
 	$center_title		= $GLOBALS['_PJ_strings']['inventory'] . ': ' . $GLOBALS['_PJ_strings']['project_list'] . " " . $customer->giveValue('customer_name');
 	include("$_PJ_root/templates/list.ihtml");
